Move ClientNodeApi impl
