+++ radix-engine/src/kernel/interpreters/mod.rs
@@ -1,9 +1,7 @@
 mod native_interpreter;
-mod scrypto_interface;
 mod scrypto_interpreter;
 mod wasm_runtime;
 
 pub use native_interpreter::*;
-pub use scrypto_interface::*;
 pub use scrypto_interpreter::*;
 pub use wasm_runtime::*;
